+++ app/gui2/src/util/__tests__/net.test.ts
@@ -1,5 +1,4 @@
-import { Err, Ok, ResultError } from '@/util/data/result'
-import { AsyncQueue, exponentialBackoff } from '@/util/net'
+import { AsyncQueue } from '@/util/net'
 import { wait } from 'lib0/promise'
 import { afterEach, beforeEach, describe, expect, test, vi } from 'vitest'
 
@@ -49,69 +48,5 @@ describe('AsyncQueue', () => {
     })
     vi.runAllTimersAsync()
     expect(await queue.waitForCompletion()).toBe(5)
-  })
-})
-
-describe('exponentialBackoff', () => {
-  test('runs successful task once', async () => {
-    const task = vi.fn(async () => Ok(1))
-    const result = await exponentialBackoff(task)
-    expect(result).toEqual({ ok: true, value: 1 })
-    expect(task).toHaveBeenCalledTimes(1)
-  })
-
-  test('retry failing task up to a limit', async () => {
-    const task = vi.fn(async () => Err(1))
-    const promise = exponentialBackoff(task, { maxRetries: 4 })
-    vi.runAllTimersAsync()
-    const result = await promise
-    expect(result).toEqual({ ok: false, error: new ResultError(1) })
-    expect(task).toHaveBeenCalledTimes(5)
-  })
-
-  test('wait before retrying', async () => {
-    const task = vi.fn(async () => Err(null))
-    exponentialBackoff(task, {
-      maxRetries: 10,
-      retryDelay: 100,
-      retryDelayMultiplier: 3,
-      retryDelayMax: 1000,
-    })
-    expect(task).toHaveBeenCalledTimes(1)
-    await vi.advanceTimersByTimeAsync(100)
-    expect(task).toHaveBeenCalledTimes(2)
-    await vi.advanceTimersByTimeAsync(300)
-    expect(task).toHaveBeenCalledTimes(3)
-    await vi.advanceTimersByTimeAsync(900)
-    expect(task).toHaveBeenCalledTimes(4)
-    await vi.advanceTimersByTimeAsync(5000)
-    expect(task).toHaveBeenCalledTimes(9)
-  })
-
-  test('retry task until success', async () => {
-    const task = vi.fn()
-    task.mockReturnValueOnce(Promise.resolve(Err(3)))
-    task.mockReturnValueOnce(Promise.resolve(Err(2)))
-    task.mockReturnValueOnce(Promise.resolve(Ok(1)))
-    const promise = exponentialBackoff(task)
-    vi.runAllTimersAsync()
-    const result = await promise
-    expect(result).toEqual({ ok: true, value: 1 })
-    expect(task).toHaveBeenCalledTimes(3)
-  })
-
-  test('call retry callback', async () => {
-    const task = vi.fn()
-    task.mockReturnValueOnce(Promise.resolve(Err(3)))
-    task.mockReturnValueOnce(Promise.resolve(Err(2)))
-    task.mockReturnValueOnce(Promise.resolve(Ok(1)))
-    const onBeforeRetry = vi.fn()
-
-    const promise = exponentialBackoff(task, { onBeforeRetry })
-    vi.runAllTimersAsync()
-    await promise
-    expect(onBeforeRetry).toHaveBeenCalledTimes(2)
-    expect(onBeforeRetry).toHaveBeenNthCalledWith(1, new ResultError(3), 0, 1000)
-    expect(onBeforeRetry).toHaveBeenNthCalledWith(2, new ResultError(2), 1, 2000)
   })
 })
